Dogfooding, trying to switch platyPS to schema 2.0.0
